web/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
